refactor: abstract away val logic
